Delcare google_api_protos (#3151)
* Delcare google_api_protos

This is kind of managed dependency and probalby needs to be declared
here

I'm going to remove it from server-infra
https://github.com/wix-private/server-infra/blob/master/framework/grpc/generator-bazel/src/main/rules/wix_scala_proto_repositories.bzl#L5

And remove load from WORKSPACES

Not sure this is the right place to define it.

Maybe it should be a separate dependnecy?
Currently I'm looking at it as part of rules_proto so added it here.

* Fix build

* Comment about @com_google_protobuf repository

GitOrigin-RevId: dc94705370cbaf9fbcad108d0ff9d8c85f2054ef